refactor: Move validate_email_not_already_in_realm.
We move this to email_validation.py.
